requirements.txt
fastapi>=0.104.0
boto3>=1.28.0
PyJWT>=2.8.0
requests>=2.31.0
pydantic>=2.0.0
python-dotenv>=1.0.0
uvicorn>=0.24.0
